+++ site/source/pages/statistiques/LazyStats.tsx
@@ -1,4 +1,4 @@
-import { Suspense, lazy } from 'react'
+import { Suspense, lazy, useState } from 'react'
 import { useTranslation } from 'react-i18next'
 
 import { TrackPage } from '@/components/ATInternetTracking'
@@ -7,8 +7,11 @@ import PageHeader from '@/components/PageHeader'
 import Privacy from '@/components/layout/Footer/Privacy'
 import Meta from '@/components/utils/Meta'
 import { ScrollToTop } from '@/components/utils/Scroll'
+import { Message } from '@/design-system'
 import { Emoji } from '@/design-system/emoji'
-import { Intro } from '@/design-system/typography/paragraphs'
+import { Spacing } from '@/design-system/layout'
+import { Switch } from '@/design-system/switch'
+import { Body, Intro } from '@/design-system/typography/paragraphs'
 
 import illustrationSvg from './illustration.svg'
 
@@ -16,6 +19,7 @@ const Stats = lazy(() => import('./Stats'))
 
 export default function StatsPage() {
 	const { t } = useTranslation()
+	const [accessibleStats, setAccessibleStats] = useState(false)
 
 	return (
 		<>
@@ -36,16 +40,29 @@ export default function StatsPage() {
 				}
 				picture={illustrationSvg}
 			>
-				{' '}
 				<Intro>
 					Découvrez nos statistiques d'utilisation mises à jour quotidiennement.
 					Les données recueillies sont anonymisées.{' '}
 					<Privacy label="En savoir plus" />
 				</Intro>
+				<Message type="info" icon>
+					<Body>
+						Cette page contient de nombreux graphiques qui ne sont pas
+						accessibles pour les lecteurs d'écran, nous avons donc mis en place
+						un mode d'accessibilité qui les affiche sous forme de tableaux.
+					</Body>
+					<Switch
+						defaultSelected={accessibleStats}
+						onChange={setAccessibleStats}
+					>
+						Activer le mode d'accessibilité sur cette page :
+					</Switch>
+					<Spacing sm />
+				</Message>
 			</PageHeader>
 
 			<Suspense fallback={<Intro>Chargement des statistiques...</Intro>}>
-				<Stats />
+				<Stats accessibleStats={accessibleStats} />
 			</Suspense>
 			<MoreInfosOnUs />
 		</>
